refactor: table package 의 table, tablegroup 을 분리하라
