+++ src/main/java/kitchenpos/table/domain/OrderTableRepository.java
@@ -1,6 +1,7 @@
 package kitchenpos.table.domain;
 
 import java.util.List;
+import kitchenpos.tableGroup.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderTableRepository extends JpaRepository<OrderTable, Long> {
