+++ src/main/java/kitchenpos/tableGroup/domain/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.domain;
+package kitchenpos.tableGroup.domain;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -9,6 +9,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.OrderTables;
 
 @Entity
 @Table(name = "table_group")
