+++ src/main/java/kitchenpos/tableGroup/domain/TableGroupRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.domain;
+package kitchenpos.tableGroup.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
