+++ src/main/java/kitchenpos/tableGroup/dto/response/TableGroupResponse.java
@@ -1,9 +1,10 @@
-package kitchenpos.table.domain.response;
+package kitchenpos.tableGroup.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.domain.response.OrderTableResponse;
+import kitchenpos.tableGroup.domain.TableGroup;
 
 public class TableGroupResponse {
