+++ src/test/java/kitchenpos/acceptance/TableGroupAcceptanceTest.java
@@ -10,9 +10,9 @@ import io.restassured.response.Response;
 import java.util.Arrays;
 import java.util.List;
 import kitchenpos.table.domain.request.OrderTableRequest;
-import kitchenpos.table.domain.request.TableGroupRequest;
+import kitchenpos.tableGroup.dto.request.TableGroupRequest;
 import kitchenpos.table.domain.response.OrderTableResponse;
-import kitchenpos.table.domain.response.TableGroupResponse;
+import kitchenpos.tableGroup.dto.response.TableGroupResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
