+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -5,7 +5,6 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -26,13 +25,13 @@ import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.application.OrderTableServiceImpl;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
-import kitchenpos.table.application.TableGroupService;
-import kitchenpos.table.domain.TableGroup;
-import kitchenpos.table.domain.TableGroupRepository;
+import kitchenpos.tableGroup.application.TableGroupService;
+import kitchenpos.tableGroup.domain.TableGroup;
+import kitchenpos.tableGroup.domain.TableGroupRepository;
 import kitchenpos.table.domain.request.OrderTableRequest;
-import kitchenpos.table.domain.request.TableGroupRequest;
+import kitchenpos.tableGroup.dto.request.TableGroupRequest;
 import kitchenpos.table.domain.response.OrderTableResponse;
-import kitchenpos.table.domain.response.TableGroupResponse;
+import kitchenpos.tableGroup.dto.response.TableGroupResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
